requirements.txt
# --- Core Dependencies ---
# Note: For CUDA 11.7, install via:
# pip install torch==1.13.1+cu117 torchvision==0.14.1+cu117 -f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1
torchvision==0.14.1

# --- OpenMMLab ---
# Note: For CUDA 11.7, install via:
# pip install mmcv-full==1.7.2 -f https://download.openmmlab.com/mmcv/dist/cu117/torch1.13/index.html
mmcv-full==1.7.2

# --- Dataset and Tools ---
lvis
matplotlib
numpy
pycocotools
scipy
six
terminaltables
cityscapesscripts
imagecorruptions
scikit-learn

# --- Development and Testing ---
pytest
yapf
flake8
isort==4.3.21
interrogate
onnx==1.7.0
onnxruntime>=1.8.0
protobuf<=3.20.1
ubelt
xdoctest>=0.10.0
